SDL does not apply heart beat timeout in ms

+++ src/components/policy/src/policy/include/policy/cache_manager.h
@@ -555,10 +555,10 @@ class CacheManager : public CacheManagerInterface {
   /**
    * Returns heart beat timeout
    * @param app_id application id
-   * @return if timeout was set then value in seconds greater zero
+   * @return if timeout was set then value in milliseconds greater zero
    * otherwise heart beat for specific application isn't set
    */
-  uint16_t HeartBeatTimeout(const std::string& app_id) const;
+  uint32_t HeartBeatTimeout(const std::string& app_id) const;
 
   /**
    * @brief Allows to generate hash from the specified string.
